Move chebpy.core.ui into chebpy.api
Previously the chebpy.core.ui methods were imported in chebpy.api, but is now moved to the user-facing part of chebpy.
The core.ui module is not used anywhere inside core and is in fact not part of the core chebpy technology.
