+++ chebpy/__init__.py
@@ -1,3 +1,3 @@
 # -*- coding: utf-8 -*-
 
-from chebpy.api import *
+from chebpy.api import chebfun, pwc
